requirements.txt
fastapi
python-multipart
torch
torchvision
pillow
facenet-pytorch
pymongo
scipy
python-dotenv
uvicorn